Moved VersionHidingCursorAdapter to adapters
